Styled Share page
added style.js
added MediaInfo.js
changed icons to svg
new CTI logos
