+++ src/pages/RadicalCollaboration/Share/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
+import useStyles from './styles.js';
+import MediaInfo from './MediaInfo.js';
 import Box from '@material-ui/core/Box';
-import Button from '@material-ui/core/Button';
 import Container from '@material-ui/core/Container';
 import Card from '@material-ui/core/Card';
 import CardActions from '@material-ui/core/CardActions';
@@ -8,87 +9,9 @@ import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import makeStyles from '@material-ui/core/styles/makeStyles';
 import { GenericHeaderSection } from '../../../components';
 import GitHubButton from 'react-github-btn';
-
-const useStyles = makeStyles((theme) => ({
-  containerStyle: {
-    paddingBottom: theme.spacing(6),
-  },
-  imageStyle: {
-    border: '1px solid',
-    borderColor: theme.palette.background.darkGray,
-    maxWidth: '100%',
-    [theme.breakpoints.down('md')]: {
-      fontSize: '36px',
-    },
-  },
-  ctiImageStyle: {
-    maxWidth: '100%',
-    [theme.breakpoints.down('xs')]: {
-      width: '102px',
-      height: '62px',
-    },
-  },
-  ctiSquareStyle: {
-    maxWidth: '100%',
-    [theme.breakpoints.down('xs')]: {
-      width: '150px',
-      height: '150px',
-    },
-  },
-  starMediaSection: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    [theme.breakpoints.down('xs')]: {
-      flexDirection: 'column',
-    },
-  },
-  socialMediaSectionStyle: {
-    padding: '60px 0 80px',
-    [theme.breakpoints.down('sm')]: {
-      padding: '47px 0 72px 0',
-    },
-  },
-  mediaInfoStyle: {
-    padding: '112px 0 153px 0 ',
-    [theme.breakpoints.down('md')]: {
-      padding: '112px 0 77px 0 ',
-    },
-    [theme.breakpoints.down('sm')]: {
-      padding: '112px 0 96px 0 ',
-    },
-  },
-  card: {
-    backgroundColor: theme.palette.spectrum.white,
-    height: '186px',
-    border: '1px solid',
-    borderRadius:'4px',
-    borderColor: theme.palette.background.darkGray,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'space-around',
-  },
-  cardTypo:{
-    textAlign:'center',
-    width:'218px',
-  },
-  media: {
-    height: '64px',
-    width: '64px',
-  },
-  tweetGrid:{
-    paddingTop:'96px',
-    paddingBottom:'32px',
-  },
-  tweetHeading:{
-    textAlign: 'center',
-    color: theme.palette.spectrum.teal,
-  },
-}));
+import { TwitterTimelineEmbed } from 'react-twitter-embed';
 
 const StarMediaSection = () => {
   const classes = useStyles();
@@ -151,11 +74,11 @@ const SocialMediaPostSection = () => {
       justify='center'
       style={{ marginTop: '24px' }}
     >
-      <Grid item>
-        <img
-          src='/images/CTI-Twitter-latest.png'
-          alt='CTI Twitter latest'
-          className={classes.imageStyle}
+      <Grid className={classes.twitterFeedStyle}>
+        <TwitterTimelineEmbed
+          sourceType="profile"
+          screenName="HackForLA"
+          options={{height: 640}}
         />
       </Grid>
     </Grid>
@@ -164,7 +87,7 @@ const SocialMediaPostSection = () => {
 const SocialMediaSection = () => {
   const classes = useStyles();
   return (
-    <Box className='containerGray'>
+    <Box>
       <Container>
         <Grid
           container
@@ -195,7 +118,6 @@ const SocialMediaSection = () => {
             alignItems='center'
             justify='space-between'
             style={{ maxWidth: '576px' }}
-            className={classes.socialMediaSectionStyle}
           >
             <a href='https://www.instagram.com/civictechindex'>
               <img src='/images/share-insta-logo.svg' alt='instagram logo' />
@@ -212,121 +134,10 @@ const SocialMediaSection = () => {
             <a href='https://github.com/civictechindex'>
               <img src='/images/share-github-logo.svg' alt='github logo' />
             </a>
-    
           </Grid>
         </Grid>
       </Container>
     </Box>
-  );
-};
-
-const MediaInfo = () => {
-  const classes = useStyles();
-  return (
-    <Container>
-      <Grid container className={classes.mediaInfoStyle}>
-        <Grid
-          container
-          direction='column'
-          alignItems='center'
-          item
-          lg={6}
-          sm={12}
-          md={12}
-          style={{ marginBottom: '32px' }}
-        >
-          <Box>
-            <Typography
-              variant='h4'
-              style={{ color: '#004364', marginBottom: '32px' }}
-            >
-              Use our Logo on your Project!
-            </Typography>
-            <Typography variant='h6' style={{ marginBottom: '16px' }}>
-              If you have done the following:
-            </Typography>
-
-            <Typography variant='body1' style={{ paddingBottom: '15px' }}>
-              • You have listed your project on Civic Tech Index
-            </Typography>
-            <Typography variant='body1' style={{ paddingBottom: '15px' }}>
-              • You have added the topic tag &quot;civictechindex&quot;to your
-              project repository
-            </Typography>
-            <Typography variant='body1' style={{ paddingBottom: '15px' }}>
-              • You have developed your own interface to search the Index
-            </Typography>
-            <Typography variant='body1' style={{ paddingBottom: '15px' }}>
-              • In articles or social media posts featuring Civic Tech Index.
-            </Typography>
-            <Typography variant='body1' style={{ paddingBottom: '15px' }}>
-              • For all other uses, please
-              <a
-                style={{
-                  color: '#0D99C6',
-                  textDecoration: ' none',
-                }}
-                href='/contact'
-              >
-                {' '}
-                contact us
-              </a>{' '}
-              for permission.
-            </Typography>
-          </Box>
-        </Grid>
-        <Grid
-          container
-          direction='column'
-          alignItems='center'
-          item
-          lg={6}
-          sm={12}
-          md={12}
-        >
-          <CtiImageSection />
-          <ButtonSection />
-        </Grid>
-      </Grid>
-    </Container>
-  );
-};
-
-const CtiImageSection = () => {
-  // const classes = useStyles();
-  return (
-    <Grid container justify='space-evenly'>
-      <img
-        src='/images/OneColor_CTI_Logo.svg'
-        alt='CTI logo'
-      />
-      <img
-        src='/images/FullColor_CTI_Logo.svg'
-        alt='CTI logo color'
-      />
-    </Grid>
-  );
-};
-
-const ButtonSection = () => {
-  return (
-    <Grid
-      container
-      direction='column'
-      alignItems='center'
-      style={{ marginTop: '48px' }}
-    >
-      <Button
-        href='https://drive.google.com/drive/folders/1EORoPWmALLXJCfHyDXJemYpO5HvnMzPN'
-        target='blank'
-      >
-        Download the CTI Assets Kit
-      </Button>
-
-      <Typography variant='body1' style={{ marginTop: '18px' }}>
-        (All sizes & formats listed above in a .zip file)
-      </Typography>
-    </Grid>
   );
 };
 
@@ -355,7 +166,7 @@ export default function ShareTheCti() {
             justify='center'
             direction='column'
             style={{
-              padding: '64px 0 47px 0',
+              padding: '0',
               textAlign: 'center',
             }}
           >
@@ -363,17 +174,15 @@ export default function ShareTheCti() {
               Tell others about us!
             </Typography>
             <Typography
-              variant='h5'
+              variant='h6'
               style={{
-                fontWeight: '400',
-                fontSize: '20px',
-                marginTop: '8px',
+                margin: '8px 0 32px',
               }}
             >
               Here are some suggested content to share with your networks.
             </Typography>
+            <StarMediaSection />
           </Grid>
-          <StarMediaSection />
           <Grid className={classes.tweetGrid}>
             <Typography
               variant='h3'
@@ -381,8 +190,8 @@ export default function ShareTheCti() {
             >
             Latest Updates from the CTI
             </Typography>
+            <SocialMediaPostSection />
           </Grid>
-          <SocialMediaPostSection />
         </Container>
         <SocialMediaSection />
       </Box>
